make: Script for running the transform
Script the XSLT process to simplify the Makefile since I'm not sure how
to make an implicit rule for `N.foo: N.bar N-1.foo`. Adds flexibility to
naming the files at the expense of running all transforms when one
changes. That expense could maybe be solved if it becomes annoying.