figures: Tweaking content structure
Ideally the minimal SVGs are expanded via the same JS code to create a
"realized" SVG in the content directory. However, node.js lacks an
XMLSerializer implementation out of the box so doing the initial update
and "redraw" into the content direction is going to require more work.
For now this will simply assume JS is enabled in the browser.